dev-agent-python/pyproject.toml
[tool.poetry]
name = "dev-agent-python"
version = "0.1.0"
description = "Production-grade Python Core for Dev-Agent (The Greater God Agent)"
authors = ["Kairos Team <engineering@kairos.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.1"
langgraph = "^0.6.4"
langchain = "^0.3.17"
langchain-openai = "^0.3.3"
faiss-cpu = "^1.7.4"
python-multipart = "^0.0.6"
requests = "^2.31.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-celery = "^0.43b0"
playwright = "^1.41.1" 
# tree-sitter = "^0.20.4"
# tree-sitter-languages = "^1.10.2"
pygls = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
